requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
async-timeout==5.0.1
    # via redis
black==24.10.0
    # via -r requirements.in
blinker==1.9.0
    # via
    #   flask
    #   sentry-sdk
cachelib==0.9.0
    # via flask-caching
cachetools==6.2.0
    # via google-auth
certifi==2025.10.5
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via
    #   black
    #   flask
coverage[toml]==7.10.7
    # via pytest-cov
deprecated==1.2.18
    # via limits
flake8==7.1.1
    # via -r requirements.in
flask==3.0.3
    # via
    #   -r requirements.in
    #   flask-caching
    #   flask-cors
    #   flask-limiter
    #   pytest-flask
    #   sentry-sdk
flask-caching==2.3.0
    # via -r requirements.in
flask-cors==5.0.0
    # via -r requirements.in
flask-limiter==3.8.0
    # via -r requirements.in
google-ai-generativelanguage==0.6.10
    # via google-generativeai
google-api-core[grpc]==2.25.2
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.184.0
    # via google-generativeai
google-auth==2.41.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.8.3
    # via -r requirements.in
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.75.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
gunicorn==23.0.0
    # via -r requirements.in
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.10
    # via requests
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
limits==5.6.0
    # via flask-limiter
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   jinja2
    #   sentry-sdk
    #   werkzeug
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
mypy==1.13.0
    # via -r requirements.in
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
ordered-set==4.1.0
    # via flask-limiter
packaging==25.0
    # via
    #   black
    #   gunicorn
    #   limits
    #   pytest
pathspec==0.12.1
    # via black
platformdirs==4.4.0
    # via black
pluggy==1.6.0
    # via pytest
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycodestyle==2.12.1
    # via flake8
pydantic==2.11.10
    # via google-generativeai
pydantic-core==2.33.2
    # via pydantic
pyflakes==3.2.0
    # via flake8
pygments==2.19.2
    # via rich
pyparsing==3.2.5
    # via httplib2
pypdf2==3.0.1
    # via -r requirements.in
pytest==8.3.3
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-flask
pytest-cov==5.0.0
    # via -r requirements.in
pytest-flask==1.3.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
redis==5.2.0
    # via -r requirements.in
requests==2.32.5
    # via google-api-core
rich==13.9.4
    # via flask-limiter
rsa==4.9.1
    # via google-auth
sentry-sdk[flask]==2.18.0
    # via -r requirements.in
tqdm==4.67.1
    # via google-generativeai
typing-extensions==4.15.0
    # via
    #   flask-limiter
    #   google-generativeai
    #   grpcio
    #   limits
    #   mypy
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
werkzeug==3.1.3
    # via
    #   flask
    #   pytest-flask
wrapt==1.17.3
    # via deprecated
